requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
python-multipart==0.0.12
pydantic==2.9.0
